+++ src/data/supplemental/schema.ts
@@ -100,6 +100,7 @@ export const TargetSelectorSchema = z.enum([
   'ACTIVE_PLAYER',
   'ALL_PLAYERS',
   'ALL_HEROES',
+  'ALL_HEROES_AND_ALLIES',
   'TRIGGERING_HERO',
   'CHOSEN_PLAYER',
   'VILLAIN',